pyproject.toml
[project]
name = "sltools"
version = "0.1.0"
description = "A collection of Slurm tools"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "rich>=14.2.0",
    "tyro>=1.0.3",
    "pyyaml>=6.0.0",
]

[project.scripts]
sltop = "sltools.sltop:_cli"
slreserve = "sltools.reserve:_cli"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
